Renamed version.py to version_helper.py otherwise it is not possible to do "from astropy.version import _get_version_py_str" from affiliated packages
